feature(discussions): adss discussion/Reply AMD module
Moves discussion/Reply constructor and prototype to a standalone AMD module
